requirements.txt
streamlit==1.40.2
agno
Pillow==10.0.0
duckduckgo-search==5.3.1
google-generativeai==0.8.3